+++ versions.tf
@@ -5,7 +5,7 @@ terraform {
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = ">= 3.0"
+      version = "~> 3.0"
     }
     # tflint-ignore: terraform_unused_required_providers
     mssql = {