+++ versions.tf
@@ -1,11 +1,10 @@
 terraform {
-  experiments      = [module_variable_optional_attrs]
-  required_version = ">= 1.0, < 1.3"
+  required_version = ">= 1.3"
 
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = "~> 3.0"
+      version = "~> 3.22"
     }
     # tflint-ignore: terraform_unused_required_providers
     mssql = {