+++ examples/main/main.tf
@@ -1,11 +1,10 @@
 terraform {
-  required_version = ">= 0.14"
-  experiments      = [module_variable_optional_attrs]
+  required_version = ">= 1.3"
 
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = "~> 2.97"
+      version = ">= 3.0"
     }
     random = {
       source  = "hashicorp/random"
@@ -13,7 +12,7 @@ terraform {
     }
     mssql = {
       source  = "betr-io/mssql"
-      version = "0.2.3"
+      version = "0.2.5"
     }
   }
 }