+++ modules/databases_users/versions.tf
@@ -4,7 +4,7 @@ terraform {
   required_providers {
     random = {
       source  = "hashicorp/random"
-      version = ">= 2.0"
+      version = ">= 3.4.3"
     }
     mssql = {
       source  = "betr-io/mssql"