+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-# v4.4.2 : 2022-09-19
+# Unreleased
+
 Changed 
-  * AZ-843: Update to ARM 3.0
+  * AZ-843: Update to AzureRM `v3.0+`
 
 # v4.4.1 - 2022-08-12