+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+Changed
+  * AZ-901: Change default value for `public_network_access_enabled` variable to `false`
+
 # v7.0.0 - 2022-09-30
 
 Breaking